Main: unify Vector classes using a template with configurable type
