+++ OgreMain/include/OgreVector2.h
@@ -1,520 +1,1 @@
-/*
------------------------------------------------------------------------------
-This source file is part of OGRE
-    (Object-oriented Graphics Rendering Engine)
-For the latest info, see http://www.ogre3d.org/
-
-Copyright (c) 2000-2014 Torus Knot Software Ltd
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
------------------------------------------------------------------------------
-*/
-#ifndef __Vector2_H__
-#define __Vector2_H__
-
-
-#include "OgrePrerequisites.h"
-#include "OgreMath.h"
-
-namespace Ogre
-{
-
-    /** \addtogroup Core
-    *  @{
-    */
-    /** \addtogroup Math
-    *  @{
-    */
-    /** Standard 2-dimensional vector.
-        @remarks
-            A direction in 2D space represented as distances along the 2
-            orthogonal axes (x, y). Note that positions, directions and
-            scaling factors can be represented by a vector, depending on how
-            you interpret the values.
-    */
-    class _OgreExport Vector2
-    {
-    public:
-        Real x, y;
-
-    public:
-        /** Default constructor.
-            @note
-                It does <b>NOT</b> initialize the vector for efficiency.
-        */
-        inline Vector2()
-        {
-        }
-
-        inline Vector2(const Real fX, const Real fY )
-            : x( fX ), y( fY )
-        {
-        }
-
-        inline explicit Vector2( const Real scaler )
-            : x( scaler), y( scaler )
-        {
-        }
-
-        inline explicit Vector2( const Real afCoordinate[2] )
-            : x( afCoordinate[0] ),
-              y( afCoordinate[1] )
-        {
-        }
-
-        inline explicit Vector2( const int afCoordinate[2] )
-        {
-            x = (Real)afCoordinate[0];
-            y = (Real)afCoordinate[1];
-        }
-
-        inline Real operator [] ( const size_t i ) const
-        {
-            assert( i < 2 );
-
-            return *(&x+i);
-        }
-
-        inline Real& operator [] ( const size_t i )
-        {
-            assert( i < 2 );
-
-            return *(&x+i);
-        }
-
-        /// Pointer accessor for direct copying
-        inline Real* ptr()
-        {
-            return &x;
-        }
-        /// Pointer accessor for direct copying
-        inline const Real* ptr() const
-        {
-            return &x;
-        }
-
-        inline bool operator == ( const Vector2& rkVector ) const
-        {
-            return ( x == rkVector.x && y == rkVector.y );
-        }
-
-        inline bool operator != ( const Vector2& rkVector ) const
-        {
-            return ( x != rkVector.x || y != rkVector.y  );
-        }
-
-        // arithmetic operations
-        inline Vector2 operator + ( const Vector2& rkVector ) const
-        {
-            return Vector2(
-                x + rkVector.x,
-                y + rkVector.y);
-        }
-
-        inline Vector2 operator - ( const Vector2& rkVector ) const
-        {
-            return Vector2(
-                x - rkVector.x,
-                y - rkVector.y);
-        }
-
-        inline Vector2 operator * ( const Real fScalar ) const
-        {
-            return Vector2(
-                x * fScalar,
-                y * fScalar);
-        }
-
-        inline Vector2 operator * ( const Vector2& rhs) const
-        {
-            return Vector2(
-                x * rhs.x,
-                y * rhs.y);
-        }
-
-        inline Vector2 operator / ( const Real fScalar ) const
-        {
-            assert( fScalar != 0.0 );
-
-            Real fInv = 1.0f / fScalar;
-
-            return Vector2(
-                x * fInv,
-                y * fInv);
-        }
-
-        inline Vector2 operator / ( const Vector2& rhs) const
-        {
-            return Vector2(
-                x / rhs.x,
-                y / rhs.y);
-        }
-
-        inline const Vector2& operator + () const
-        {
-            return *this;
-        }
-
-        inline Vector2 operator - () const
-        {
-            return Vector2(-x, -y);
-        }
-
-        // overloaded operators to help Vector2
-        inline friend Vector2 operator * ( const Real fScalar, const Vector2& rkVector )
-        {
-            return Vector2(
-                fScalar * rkVector.x,
-                fScalar * rkVector.y);
-        }
-
-        inline friend Vector2 operator / ( const Real fScalar, const Vector2& rkVector )
-        {
-            return Vector2(
-                fScalar / rkVector.x,
-                fScalar / rkVector.y);
-        }
-
-        inline friend Vector2 operator + (const Vector2& lhs, const Real rhs)
-        {
-            return Vector2(
-                lhs.x + rhs,
-                lhs.y + rhs);
-        }
-
-        inline friend Vector2 operator + (const Real lhs, const Vector2& rhs)
-        {
-            return Vector2(
-                lhs + rhs.x,
-                lhs + rhs.y);
-        }
-
-        inline friend Vector2 operator - (const Vector2& lhs, const Real rhs)
-        {
-            return Vector2(
-                lhs.x - rhs,
-                lhs.y - rhs);
-        }
-
-        inline friend Vector2 operator - (const Real lhs, const Vector2& rhs)
-        {
-            return Vector2(
-                lhs - rhs.x,
-                lhs - rhs.y);
-        }
-
-        // arithmetic updates
-        inline Vector2& operator += ( const Vector2& rkVector )
-        {
-            x += rkVector.x;
-            y += rkVector.y;
-
-            return *this;
-        }
-
-        inline Vector2& operator += ( const Real fScaler )
-        {
-            x += fScaler;
-            y += fScaler;
-
-            return *this;
-        }
-
-        inline Vector2& operator -= ( const Vector2& rkVector )
-        {
-            x -= rkVector.x;
-            y -= rkVector.y;
-
-            return *this;
-        }
-
-        inline Vector2& operator -= ( const Real fScaler )
-        {
-            x -= fScaler;
-            y -= fScaler;
-
-            return *this;
-        }
-
-        inline Vector2& operator *= ( const Real fScalar )
-        {
-            x *= fScalar;
-            y *= fScalar;
-
-            return *this;
-        }
-
-        inline Vector2& operator *= ( const Vector2& rkVector )
-        {
-            x *= rkVector.x;
-            y *= rkVector.y;
-
-            return *this;
-        }
-
-        inline Vector2& operator /= ( const Real fScalar )
-        {
-            assert( fScalar != 0.0 );
-
-            Real fInv = 1.0f / fScalar;
-
-            x *= fInv;
-            y *= fInv;
-
-            return *this;
-        }
-
-        inline Vector2& operator /= ( const Vector2& rkVector )
-        {
-            x /= rkVector.x;
-            y /= rkVector.y;
-
-            return *this;
-        }
-
-        /// @copydoc Vector3::length
-        inline Real length () const
-        {
-            return Math::Sqrt( x * x + y * y );
-        }
-
-        /// @copydoc Vector3::squaredLength
-        inline Real squaredLength () const
-        {
-            return x * x + y * y;
-        }
-
-        /// @copydoc Vector3::distance
-        inline Real distance(const Vector2& rhs) const
-        {
-            return (*this - rhs).length();
-        }
-
-        /// @copydoc Vector3::squaredDistance
-        inline Real squaredDistance(const Vector2& rhs) const
-        {
-            return (*this - rhs).squaredLength();
-        }
-
-        /// @copydoc Vector3::dotProduct
-        inline Real dotProduct(const Vector2& vec) const
-        {
-            return x * vec.x + y * vec.y;
-        }
-
-        /// @copydoc Vector3::normalise
-        inline Real normalise()
-        {
-            Real fLength = Math::Sqrt( x * x + y * y);
-
-            // Will also work for zero-sized vectors, but will change nothing
-            // We're not using epsilons because we don't need to.
-            // Read http://www.ogre3d.org/forums/viewtopic.php?f=4&t=61259
-            if ( fLength > Real(0.0f) )
-            {
-                Real fInvLength = 1.0f / fLength;
-                x *= fInvLength;
-                y *= fInvLength;
-            }
-
-            return fLength;
-        }
-
-        /** Returns a vector at a point half way between this and the passed
-            in vector.
-        */
-        inline Vector2 midPoint( const Vector2& vec ) const
-        {
-            return Vector2(
-                ( x + vec.x ) * 0.5f,
-                ( y + vec.y ) * 0.5f );
-        }
-
-        /** Returns true if the vector's scalar components are all greater
-            that the ones of the vector it is compared against.
-        */
-        inline bool operator < ( const Vector2& rhs ) const
-        {
-            if( x < rhs.x && y < rhs.y )
-                return true;
-            return false;
-        }
-
-        /** Returns true if the vector's scalar components are all smaller
-            that the ones of the vector it is compared against.
-        */
-        inline bool operator > ( const Vector2& rhs ) const
-        {
-            if( x > rhs.x && y > rhs.y )
-                return true;
-            return false;
-        }
-
-        /** Sets this vector's components to the minimum of its own and the
-            ones of the passed in vector.
-            @remarks
-                'Minimum' in this case means the combination of the lowest
-                value of x, y and z from both vectors. Lowest is taken just
-                numerically, not magnitude, so -1 < 0.
-        */
-        inline void makeFloor( const Vector2& cmp )
-        {
-            if( cmp.x < x ) x = cmp.x;
-            if( cmp.y < y ) y = cmp.y;
-        }
-
-        /** Sets this vector's components to the maximum of its own and the
-            ones of the passed in vector.
-            @remarks
-                'Maximum' in this case means the combination of the highest
-                value of x, y and z from both vectors. Highest is taken just
-                numerically, not magnitude, so 1 > -3.
-        */
-        inline void makeCeil( const Vector2& cmp )
-        {
-            if( cmp.x > x ) x = cmp.x;
-            if( cmp.y > y ) y = cmp.y;
-        }
-
-        /** Generates a vector perpendicular to this vector (eg an 'up' vector).
-            @remarks
-                This method will return a vector which is perpendicular to this
-                vector. There are an infinite number of possibilities but this
-                method will guarantee to generate one of them. If you need more
-                control you should use the Quaternion class.
-        */
-        inline Vector2 perpendicular(void) const
-        {
-            return Vector2 (-y, x);
-        }
-
-        /** Calculates the 2 dimensional cross-product of 2 vectors, which results
-            in a single floating point value which is 2 times the area of the triangle.
-        */
-        inline Real crossProduct( const Vector2& rkVector ) const
-        {
-            return x * rkVector.y - y * rkVector.x;
-        }
-
-        /** Generates a new random vector which deviates from this vector by a
-            given angle in a random direction.
-            @remarks
-                This method assumes that the random number generator has already
-                been seeded appropriately.
-            @param angle
-                The angle at which to deviate in radians
-            @return
-                A random vector which deviates from this vector by angle. This
-                vector will not be normalised, normalise it if you wish
-                afterwards.
-        */
-        inline Vector2 randomDeviant(Radian angle) const
-        {
-            angle *= Math::RangeRandom(-1, 1);
-            Real cosa = Math::Cos(angle);
-            Real sina = Math::Sin(angle);
-            return Vector2(cosa * x - sina * y,
-                           sina * x + cosa * y);
-        }
-
-        /** Returns true if this vector is zero length. */
-        inline bool isZeroLength(void) const
-        {
-            Real sqlen = (x * x) + (y * y);
-            return (sqlen < (1e-06 * 1e-06));
-
-        }
-
-        /** As normalise, except that this vector is unaffected and the
-            normalised vector is returned as a copy. */
-        inline Vector2 normalisedCopy(void) const
-        {
-            Vector2 ret = *this;
-            ret.normalise();
-            return ret;
-        }
-
-        /** Calculates a reflection vector to the plane with the given normal .
-        @remarks NB assumes 'this' is pointing AWAY FROM the plane, invert if it is not.
-        */
-        inline Vector2 reflect(const Vector2& normal) const
-        {
-            return Vector2( *this - ( 2 * this->dotProduct(normal) * normal ) );
-        }
-
-        /// Check whether this vector contains valid values
-        inline bool isNaN() const
-        {
-            return Math::isNaN(x) || Math::isNaN(y);
-        }
-
-        /**  Gets the angle between 2 vectors.
-        @remarks
-            Vectors do not have to be unit-length but must represent directions.
-        */
-        inline Ogre::Radian angleBetween(const Ogre::Vector2& other) const
-        {       
-            Ogre::Real lenProduct = length() * other.length();
-            // Divide by zero check
-            if(lenProduct < 1e-6f)
-                lenProduct = 1e-6f;
-        
-            Ogre::Real f = dotProduct(other) / lenProduct;
-    
-            f = Ogre::Math::Clamp(f, (Ogre::Real)-1.0, (Ogre::Real)1.0);
-            return Ogre::Math::ACos(f);
-        }
-
-        /**  Gets the oriented angle between 2 vectors.
-        @remarks
-            Vectors do not have to be unit-length but must represent directions.
-            The angle is comprised between 0 and 2 PI.
-        */
-        inline Ogre::Radian angleTo(const Ogre::Vector2& other) const
-        {
-            Ogre::Radian angle = angleBetween(other);
-        
-            if (crossProduct(other)<0)          
-                angle = (Ogre::Radian)Ogre::Math::TWO_PI - angle;       
-
-            return angle;
-        }
-
-        // special points
-        static const Vector2 ZERO;
-        static const Vector2 UNIT_X;
-        static const Vector2 UNIT_Y;
-        static const Vector2 NEGATIVE_UNIT_X;
-        static const Vector2 NEGATIVE_UNIT_Y;
-        static const Vector2 UNIT_SCALE;
-
-        /** Function for writing to a stream.
-        */
-        inline friend std::ostream& operator <<
-            ( std::ostream& o, const Vector2& v )
-        {
-            o << "Vector2(" << v.x << ", " << v.y <<  ")";
-            return o;
-        }
-    };
-    /** @} */
-    /** @} */
-
-}
-#endif
+#include "OgreVector.h"
